Shade and upgrade Guava
Guava regularly bumps its major version and includes method removals in the process of doing so, which fits with Semantic Versioning, but which limits reuse of libraries that externally depend on it. This removes that obstacle by shading/removing Guava as an external dependency, so it can be upgraded locally without affecting others.

Also updates Apache Commons libraries (do not need shading as they have a much longer maintenance cycle), and Slf4J (Does not need shading as it has a very reputable history in terms of backwards compatibility). Logback is only used as a test dependency, so bumping to keep up with latest bug fixes, but no effect on downstream projects.

Fixes one removed method since Guava-18.0 by replacing it with a standard Java Collections API method.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>